§ LLM api request functions merged into 1
