+++ front/src/components/Chat/Prompt.jsx
@@ -13,7 +13,11 @@ import upload from "../../assets/add.svg";
 import mic from "../../assets/icon_mic.svg";
 import stop from "../../assets/stop_listening.svg";
 import pause from "../../assets/pause.svg";
-import { cancelRequest, fetchLLMResponse, updateMemory } from "../../apis/LlmRequestApi";
+import {
+  cancelRequest,
+  fetchLLMResponse,
+  updateMemory,
+} from "../../apis/LlmRequestApi";
 import Tooltip from "../Others/Tooltip";
 import { Trans, useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
@@ -29,6 +33,7 @@ import {
   selectAllMemories,
   deleteAllMemories,
 } from "../../Redux/reducers/userMemorySlice";
+import handleLLMResponse from "../../utils/handleLLMResponse";
 
 //Variable
 const languageMap = {
@@ -87,167 +92,27 @@ function Prompt({
   };
 
   // Main function to fetch and process LLM response
-  async function getRes(updatedConversation) {
-    dispatch(setIsResponding(true));
-    setLoading(true);
-
-    // Check if selected model supports image input
-    const imageSupport = modelList.some(
-      (modelX) =>
-        modelX.name === localState.settings["model-name"] &&
-        modelX.input.includes("image")
-    );
-    // Check if selected model supports video input
-    const videoSupport = modelList.some(
-      (modelX) =>
-        modelX.name === localState.settings["model-name"] &&
-        modelX.input.includes("video")
-    );
-
-    // Process conversation based on image/video support
-    let processedConversation = updatedConversation;
-    if (!imageSupport && !videoSupport) {
-      // Remove image content if model doesn't support images/videos
-      processedConversation = updatedConversation.map((message) => {
-        if (message.role === "user" && Array.isArray(message.content)) {
-          return {
-            role: "user",
-            content: message.content
-              .filter((item) => item.type === "text")
-              .map((item) => item.text)
-              .join("\n"),
-          };
-        }
-        return message;
-      });
-    }
-
-    // Filter out info messages only for API call
-    const conversationForAPI = processedConversation.filter(
-      (message) => message.role !== "info"
-    );
-
-    // Prepare system prompt with memory if enabled
-    let finalSystemPrompt = localState.settings.systemPrompt;
-
-    if (localState.settings.memory >= 1 && memories.length > 0) {
-      const memoryContext = memories.map((memory) => memory.text).join("\n");
-      const memorySection = `\n\n--- User Memory ---\nThe following information represents the user's preferences, important details, and context from previous conversations. Use this information when relevant to provide a more personalized and contextual response:\n\n${memoryContext}\n--- End User Memory ---`;
-      finalSystemPrompt = finalSystemPrompt + memorySection;
-    }
-
-    if (selectedFiles.length > 0) {
-      const imageFiles = selectedFiles.filter((file) => file.type === "image");
-      const videoFiles = selectedFiles.filter((file) => file.type === "video");
-      const textFiles = selectedFiles.filter(
-        (file) => file.type !== "image" && file.type !== "video"
-      );
-
-      const imageContent = imageFiles.map((imageFile) => ({
-        type: "image_url",
-        image_url: {
-          url: imageFile.text,
-        },
-      }));
-
-      const videoContent = videoFiles.map(() => ({
-        type: "video_url",
-        video_url: {
-          url: "",
-        },
-      }));
-
-      const textContent = textFiles.map((file) => ({
-        name: file.name,
-        fileType: "text",
-        content:
-          file.fileType === "pdf"
-            ? `${file.name}: ${file.processedContent}`
-            : `${file.name}: ${file.content}`,
-        type: "text",
-        size: file.size,
-      }));
-
-      setLocalState((prevState) => ({
-        ...prevState,
-        responses: [
-          ...prevState.responses,
-          {
-            prompt: prevState.prompt,
-            images: imageContent,
-            videos: videoContent,
-            textFiles: textContent,
-            response: "",
-          },
-        ],
-      }));
-    } else {
-      // Add response entry without images
-      setLocalState((prevState) => ({
-        ...prevState,
-        responses: [
-          ...prevState.responses,
-          {
-            prompt: prevState.prompt,
-            response: "",
-          },
-        ],
-      }));
-    }
-
-    // Clear prompt after processing
-    updateLocalState({ prompt: "" });
-
-    try {
-      const response = await fetchLLMResponse(
-        processedConversation,
-        finalSystemPrompt,
-        localState.settings.model,
-        localState.settings.temperature,
-        localState.settings.top_p,
-        localState.arcana,
-        setLocalState,
-        setShowModalSession,
-        setShowBadRequest,
-        conversationForAPI,
-        isArcanaSupported
-      );
-
-      try {
-        // TODO Update memory when necessary
-        if (localState.settings.memory >= 2) {
-          const response = await updateMemory(processedConversation, memories);
-          const cleanedResponse = response.replace(/,(\s*[}$])/g, "$1");
-          const jsonResponse = JSON.parse(cleanedResponse);
-          if (jsonResponse.store) {
-            const memoryText = jsonResponse.memory_sentence.trim();
-            dispatch(addMemory({ text: memoryText }));
-            console.log ("New memory:", memoryText);
-          }
-        }
-      } catch (error) {
-        console.error("Failed to update memory: ", error.name, error.message)
-      }
- 
-      dispatch(setIsResponding(false));
-      setLoading(false);
-      setSelectedFiles([]);
-    } catch (error) {
-      dispatch(setIsResponding(false));
-      setLoading(false);
-      setSelectedFiles([]);
-
-      // Handle different error types
-      if (error.name === "AbortError") {
-        notifyError("Request aborted.");
-      } else if (error.message) {
-        notifyError(error.message);
-      } else {
-        notifyError("An unknown error occurred");
-      }
-    }
-  }
-
+  const getRes = async (updatedConversation) => {
+    await handleLLMResponse({
+      operationType: "new",
+      updatedConversation,
+      dispatch,
+      localState,
+      updateLocalState,
+      setLocalState,
+      setLoading,
+      selectedFiles,
+      setSelectedFiles,
+      modelList,
+      memories,
+      isArcanaSupported,
+      updateMemory,
+      fetchLLMResponse,
+      notifyError,
+      setShowModalSession,
+      setShowBadRequest,
+    });
+  };
   // Convert file size from bytes to human-readable format (e.g., KB, MB, GB)
   function formatFileSize(bytes) {
     const units = ["Bytes", "KB", "MB", "GB", "TB"];
